improve typechecking for pod types and add ExtPodKind
