+++ frontend/packages/console-shared/src/utils/pod-utils.ts
@@ -7,7 +7,7 @@ import {
 } from '@console/internal/module/k8s';
 import { checkAccess } from '@console/internal/components/utils';
 import { podColor } from '../constants/pod';
-import { Pod } from '../types/pod';
+import { ExtPodKind } from '../types/pod';
 import { DEPLOYMENT_STRATEGY, DEPLOYMENT_PHASE } from '../constants';
 import { PodControllerOverviewItem, DeploymentStrategy } from '../types';
 
@@ -149,11 +149,11 @@ export const isIdled = (deploymentConfig: K8sResourceKind): boolean => {
 
 export const getPodData = (
   dc: K8sResourceKind,
-  pods: Pod[],
+  pods: ExtPodKind[],
   current: PodControllerOverviewItem,
   previous: PodControllerOverviewItem,
   isRollingOut: boolean,
-): { inProgressDeploymentData: Pod[] | null; completedDeploymentData: Pod[] } => {
+): { inProgressDeploymentData: ExtPodKind[] | null; completedDeploymentData: ExtPodKind[] } => {
   const strategy: DeploymentStrategy = _.get(dc, ['spec', 'strategy', 'type'], null);
   const currentDeploymentphase = current && current.phase;
   const currentPods = current && current.pods;
